requirements.txt
Flask==0.9
Flask-Babel==0.8
Flask-Login==0.2.9
Flask-Mail==0.7.6
Flask-OpenID==1.2.1
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.4
Flask-WhooshAlchemy==0.54a
Jinja2==2.7.2
MarkupSafe==0.18
MySQL-python==1.2.5
SQLAlchemy==0.7.9
WTForms==1.0.5
Werkzeug==0.9.4
argparse==1.2.1
ecdsa==0.10
flup==1.0.2
paramiko==1.12.2
pycrypto==2.6.1
pypayex==0.4.5
python-openid==2.2.5
pytz==2013b
sqlalchemy-migrate==0.7.2
suds==0.4
wsgiref==0.1.2

